Restructured code
moved logic from RestAPI to AppManager and OpenShiftManager
