+++ src/main/java/org/project/openbaton/nubomedia/paas/api/RestAPI.java
@@ -30,7 +30,6 @@ import org.project.openbaton.nubomedia.paas.exceptions.openshift.NameStructureEx
 import org.project.openbaton.nubomedia.paas.exceptions.openshift.UnauthorizedException;
 import org.project.openbaton.nubomedia.paas.messages.*;
 import org.project.openbaton.nubomedia.paas.model.persistence.Application;
-import org.project.openbaton.nubomedia.paas.repository.application.ApplicationRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -38,11 +37,9 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.client.ResourceAccessException;
 
 import javax.annotation.PostConstruct;
 import javax.validation.Valid;
-import java.util.ArrayList;
 import java.util.List;
 
 
@@ -59,8 +56,6 @@ public class RestAPI {
     private OpenShiftManager osmanager;
     @Autowired
     private OpenbatonManager obmanager;
-    @Autowired
-    private ApplicationRepository appRepo;
 
     @Autowired
     private AppManager appManager;
@@ -80,7 +75,6 @@ public class RestAPI {
     }
 
     /**
-     *
      * @param request
      * @param projectId
      * @return
@@ -92,38 +86,14 @@ public class RestAPI {
      * @throws StunServerException
      */
     @RequestMapping(value = "/app", method = RequestMethod.POST)
-    public
+
     @ResponseBody
-    NubomediaCreateAppResponse createApp(@RequestBody NubomediaCreateAppRequest request, @RequestHeader(value = "project-id") String projectId) throws SDKException, UnauthorizedException, DuplicatedException, NameStructureException, turnServerException, StunServerException {
-        if (token == null) {
-            throw new UnauthorizedException("No auth-token header");
-        }
-        if (request.getAppName().length() > 18) {
-            throw new NameStructureException("Name is too long");
-        }
-        if (request.getAppName().contains(".")) {
-            throw new NameStructureException("Name can't contains dots");
-        }
-
-        if (request.getAppName().contains("_")) {
-            throw new NameStructureException("Name can't contain underscore");
-        }
-
-        if (!request.getAppName().matches("[a-z0-9]+(?:[._-][a-z0-9]+)*")) {
-            throw new NameStructureException("Name must match [a-z0-9]+(?:[._-][a-z0-9]+)*");
-        }
-
-        if (!appRepo.findByAppName(request.getAppName()).isEmpty()) {
-            throw new DuplicatedException("Application with " + request.getAppName() + " already exist");
-        }
-
-        logger.debug("REQUEST" + request.toString());
+    public NubomediaCreateAppResponse createApp(@RequestBody NubomediaCreateAppRequest request, @RequestHeader(value = "project-id") String projectId) throws SDKException, UnauthorizedException, DuplicatedException, NameStructureException, turnServerException, StunServerException {
         Application app = appManager.createApplication(request, projectId, token);
         return new NubomediaCreateAppResponse(app, 200);
     }
 
     /**
-     *
      * @param id
      * @param projectId
      * @return
@@ -131,20 +101,16 @@ public class RestAPI {
      * @throws UnauthorizedException
      */
     @RequestMapping(value = "/app/{id}", method = RequestMethod.GET)
-    public
+
     @ResponseBody
-    Application getApp(@PathVariable("id") String id, @RequestHeader(value = "project-id") String projectId) throws ApplicationNotFoundException, UnauthorizedException {
+    public Application getApp(@PathVariable("id") String id, @RequestHeader(value = "project-id") String projectId) throws ApplicationNotFoundException, UnauthorizedException {
         logger.info("Request status for " + id + " app");
-        Application app = null;
-        if ( (app = appRepo.findFirstByAppIdAndProjectId(id, projectId)) == null) {
-            throw new ApplicationNotFoundException("Application with ID in Project " + projectId + " not found");
-        }
+        Application app = appManager.getApp(projectId, id);
         return app;
 
     }
 
     /**
-     *
      * @param projectId
      * @return
      * @throws UnauthorizedException
@@ -153,78 +119,29 @@ public class RestAPI {
     @RequestMapping(value = "/app", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
     public Iterable<Application> getApps(@RequestHeader(value = "project-id") String projectId) throws UnauthorizedException, ApplicationNotFoundException {
         logger.debug("Received request GET Applications");
-        Iterable<Application> applications = this.appRepo.findByProjectId(projectId);
+        Iterable<Application> applications = appManager.getApps(projectId);
         logger.debug("Returning from GET Applications " + applications);
         return applications;
     }
 
 
     /**
-     *
      * @param id
      * @param projectId
      * @return
      * @throws UnauthorizedException
      */
     @RequestMapping(value = "/app/{id}", method = RequestMethod.DELETE)
-    public
+
     @ResponseBody
-    NubomediaDeleteAppResponse deleteApp(@PathVariable("id") String id, @RequestHeader(value = "project-id") String projectId) throws UnauthorizedException {
-        logger.debug("Received delete Application (id " + id+")request");
-        Application app = null;
-        if ((app=appRepo.findFirstByAppIdAndProjectId(id, projectId)) == null) {
-            return new NubomediaDeleteAppResponse(id, "Application not found in Project " + projectId + " not found", "null", 404);
-        }
-
-        logger.debug("Retrieved Application to be deleted " + app);
-
-        // check that the app has been instantiated on openshift
-        if (!app.isResourceOK()) {
-
-            String name = app.getName();
-            String projectName = app.getProjectName();
-
-            checkAppStatus(app);
-
-            appRepo.delete(app);
-            return new NubomediaDeleteAppResponse(id, name, projectName, 200);
-
-        }
-
-        if (app.getStatus().equals(AppStatus.PAAS_RESOURCE_MISSING)) {
-            obmanager.deleteDescriptor(app.getMediaServerGroup().getNsdID());
-            obmanager.deleteRecord(app.getMediaServerGroup().getNsrID());
-            appRepo.delete(app);
-
-            return new NubomediaDeleteAppResponse(id, app.getName(), app.getProjectName(), 200);
-        }
-
-
-        obmanager.deleteRecord(app.getMediaServerGroup().getNsrID());
-        HttpStatus resDelete = HttpStatus.BAD_REQUEST;
-        try {
-            resDelete = osmanager.deleteApplication(token, app.getName());
-        } catch (ResourceAccessException e) {
-            logger.info("Application does not exist on the PaaS");
-        }
-
-        appRepo.delete(app);
-
-        return new NubomediaDeleteAppResponse(id, app.getName(), app.getProjectName(), resDelete.value());
-    }
-
-    private void checkAppStatus(Application app) {
-        if (app.getStatus().equals(AppStatus.CREATED) || app.getStatus().equals(AppStatus.INITIALIZING) || app.getStatus().equals(AppStatus.FAILED)) {
-            logger.debug("media server group: " + app.getMediaServerGroup());
-            obmanager.deleteDescriptor(app.getMediaServerGroup().getNsdID());
-            if (!app.getStatus().equals(AppStatus.FAILED) && obmanager.existRecord(app.getMediaServerGroup().getNsrID())) {
-                obmanager.deleteRecord(app.getMediaServerGroup().getNsrID());
-            }
-        }
+    public NubomediaDeleteAppResponse deleteApp(@PathVariable("id") String id, @RequestHeader(value = "project-id") String projectId) throws UnauthorizedException {
+        logger.debug("Received delete Application (id " + id + ")request");
+        return appManager.deleteApp(projectId, id);
     }
 
     /**
      * Deletes all the Apps specified in the list of ids
+     *
      * @param projectId
      * @param ids
      * @throws NotFoundException
@@ -235,67 +152,23 @@ public class RestAPI {
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void multipleDelete(@RequestHeader(value = "project-id") String projectId, @RequestBody @Valid List<String> ids) throws NotFoundException, NotAllowedException, UnauthorizedException {
         for (String id : ids)
-            deleteApp(id,projectId);
+            deleteApp(id, projectId);
     }
 
     /**
-     *
      * @param projectId
      * @return
      * @throws UnauthorizedException
      */
     @RequestMapping(value = "/app", method = RequestMethod.DELETE)
-    public
+
     @ResponseBody
-    NubomediaDeleteAppsProjectResponse deleteApp(@RequestHeader(value = "project-id") String projectId) throws UnauthorizedException {
-
-        if (appRepo.findByProjectId(projectId) == null || appRepo.findByProjectId(projectId).size() == 0) {
-            return new NubomediaDeleteAppsProjectResponse(projectId, "Not Found any Applications in this project", null, 404);
-        }
-
-        NubomediaDeleteAppsProjectResponse response = new NubomediaDeleteAppsProjectResponse(projectId, "", new ArrayList<NubomediaDeleteAppResponse>(), 200);
-
-        List<Application> apps = appRepo.findByProjectId(projectId);
-
-        logger.debug("Deleting " + apps);
-        for (Application app : apps) {
-            if (!app.isResourceOK()) {
-                checkAppStatus(app);
-                appRepo.delete(app);
-                response.getResponses().add(new NubomediaDeleteAppResponse(app.getId(), app.getName(), app.getProjectName(), 200));
-                break;
-            }
-
-            if (app.getStatus().equals(AppStatus.PAAS_RESOURCE_MISSING)) {
-                obmanager.deleteRecord(app.getMediaServerGroup().getNsrID());
-                appRepo.delete(app);
-                response.getResponses().add(new NubomediaDeleteAppResponse(app.getId(), app.getName(), app.getProjectName(), 200));
-                break;
-            }
-            obmanager.deleteRecord(app.getMediaServerGroup().getNsrID());
-            HttpStatus resDelete = HttpStatus.BAD_REQUEST;
-            try {
-                resDelete = osmanager.deleteApplication(token, app.getName());
-            } catch (ResourceAccessException e) {
-                logger.info("PaaS Missing");
-            }
-
-            appRepo.delete(app);
-            response.getResponses().add(new NubomediaDeleteAppResponse(app.getId(), app.getName(), app.getProjectName(), resDelete.value()));
-        }
-        for (NubomediaDeleteAppResponse singleRes : response.getResponses()) {
-            if (singleRes.getCode() != 200) {
-                response.setCode(singleRes.getCode());
-                response.setMessage("Not all applications were deleted successfully");
-                return response;
-            }
-        }
-        response.setMessage("All applications of this project were removed successfully");
-        return response;
+    public NubomediaDeleteAppsProjectResponse deleteApps(@RequestHeader(value = "project-id") String projectId) throws UnauthorizedException {
+        logger.info("Requested delete of all Apps of project " + projectId);
+        return appManager.deleteApps(projectId);
     }
 
     /**
-     *
      * @param id
      * @param projectId
      * @return
@@ -303,62 +176,18 @@ public class RestAPI {
      * @throws ApplicationNotFoundException
      */
     @RequestMapping(value = "/app/{id}/buildlogs", method = RequestMethod.GET)
-    public
     @ResponseBody
-    NubomediaBuildLogs getBuildLogs(@PathVariable("id") String id, @RequestHeader(value = "project-id") String projectId) throws UnauthorizedException, ApplicationNotFoundException {
-
+    public NubomediaBuildLogs getBuildLogs(@PathVariable("id") String id, @RequestHeader(value = "project-id") String projectId) throws UnauthorizedException, ApplicationNotFoundException {
+        logger.debug("Requested build log of Application " + id + " of project " + projectId);
         if (token == null) {
             throw new UnauthorizedException("no auth-token header");
         }
-
-        NubomediaBuildLogs res = new NubomediaBuildLogs();
-
-        if (appRepo.findFirstByAppIdAndProjectId(id, projectId) == null) {
-            throw new ApplicationNotFoundException("Application with ID in Project " + projectId + " not found");
-        }
-
-        Application app = appRepo.findFirstByAppIdAndProjectId(id, projectId);
-
-        if (app.getStatus().equals(AppStatus.FAILED) && !app.isResourceOK()) {
-
-            res.setId(id);
-            res.setAppName(app.getName());
-            res.setProjectName(app.getProjectName());
-            res.setLog("Something wrong on retrieving resources");
-
-        } else if (app.getStatus().equals(AppStatus.CREATED) || app.getStatus().equals(AppStatus.INITIALIZING)) {
-            res.setId(id);
-            res.setAppName(app.getName());
-            res.setProjectName(app.getProjectName());
-            res.setLog("The application is retrieving resources " + app.getStatus());
-
-            return res;
-        } else if (app.getStatus().equals(AppStatus.PAAS_RESOURCE_MISSING)) {
-            res.setId(id);
-            res.setAppName(app.getName());
-            res.setProjectName(app.getProjectName());
-            res.setLog("PaaS components are missing, send an email to the administrator to check the PaaS status");
-
-            return res;
-        } else {
-
-            res.setId(id);
-            res.setAppName(app.getName());
-            res.setProjectName(app.getProjectName());
-            try {
-                res.setLog(osmanager.getBuildLogs(token, app.getName()));
-            } catch (ResourceAccessException e) {
-                app.setStatus(AppStatus.PAAS_RESOURCE_MISSING);
-                appRepo.save(app);
-                res.setLog("Openshift is not responding, app " + app.getName() + " is not anymore available");
-            }
-        }
-
-        return res;
+        return appManager.getBuildLogs(projectId, id);
     }
 
     /**
      * s
+     *
      * @param id
      * @param podName
      * @param projectId
@@ -367,60 +196,39 @@ public class RestAPI {
      * @throws ApplicationNotFoundException
      */
     @RequestMapping(value = "/app/{id}/logs/{podName}", method = RequestMethod.GET)
-    public
     @ResponseBody
-    String getApplicationLogs(@PathVariable("id") String id, @PathVariable("podName") String podName, @RequestHeader(value = "project-id") String projectId) throws UnauthorizedException, ApplicationNotFoundException {
-
+    public String getApplicationLogs(@PathVariable("id") String id, @PathVariable("podName") String podName, @RequestHeader(value = "project-id") String projectId) throws UnauthorizedException, ApplicationNotFoundException {
+        logger.debug("Requested Application logs of Application " + id + " of pod " + podName + " of project " + projectId);
         if (token == null) {
             throw new UnauthorizedException("no auth-token header");
         }
-
-        if (appRepo.findFirstByAppIdAndProjectId(id, projectId) == null) {
-            throw new ApplicationNotFoundException("Application with ID in Project " + projectId + " not found");
-        }
-
-        Application app = appRepo.findFirstByAppIdAndProjectId(id, projectId);
-
-        if (!app.getStatus().equals(AppStatus.RUNNING)) {
-            return "Application Status " + app.getStatus() + ", logs are not available until the status is RUNNING";
-        }
-
-        return osmanager.getApplicationLog(token, app.getName(), podName);
-
+        return appManager.getApplicationLogs(projectId, id, podName);
     }
 
-
     @RequestMapping(value = "/secret", method = RequestMethod.POST)
-    public
     @ResponseBody
-    String createSecret(@RequestBody NubomediaCreateSecretRequest ncsr, @RequestHeader(value = "project-id") String projectId) throws UnauthorizedException {
-
+    public String createSecret(@RequestBody NubomediaCreateSecretRequest ncsr, @RequestHeader(value = "project-id") String projectId) throws UnauthorizedException {
+        logger.debug("Requested creation of new secret for " + ncsr.getProjectName() + " with key " + ncsr.getPrivateKey());
         if (token == null) {
             throw new UnauthorizedException("no auth-token header");
         }
-
-        logger.debug("Creating new secret for " + ncsr.getProjectName() + " with key " + ncsr.getPrivateKey());
         return osmanager.createSecret(token, ncsr.getPrivateKey());
     }
 
     @RequestMapping(value = "/secret/{projectName}/{secretName}", method = RequestMethod.DELETE)
-    public
     @ResponseBody
-    NubomediaDeleteSecretResponse deleteSecret(@PathVariable("secretName") String secretName, @PathVariable("projectName") String projectName) throws UnauthorizedException {
-
+    public NubomediaDeleteSecretResponse deleteSecret(@PathVariable("secretName") String secretName, @PathVariable("projectName") String projectName) throws UnauthorizedException {
+        logger.debug("Requested deletion of secret " + secretName);
         if (token == null) {
             throw new UnauthorizedException("no auth-token header");
         }
-
         HttpStatus deleteStatus = osmanager.deleteSecret(token, secretName);
         return new NubomediaDeleteSecretResponse(secretName, projectName, deleteStatus.value());
     }
 
     @RequestMapping(value = "/auth", method = RequestMethod.POST)
-    public
     @ResponseBody
-    NubomediaAuthorizationResponse authorize(@RequestBody NubomediaAuthorizationRequest request) throws UnauthorizedException {
-
+    public NubomediaAuthorizationResponse authorize(@RequestBody NubomediaAuthorizationRequest request) throws UnauthorizedException {
         String token = osmanager.authenticate(request.getUsername(), request.getPassword());
         switch (token) {
             case "Unauthorized":
